+++ main.c
@@ -138,6 +138,8 @@ main_usage(void)
 "  -O          deny all OCSP requests on all proxyspecs\n"
 "  -P          passthrough SSL connections if they cannot be split because of\n"
 "              client cert auth or no matching cert and no CA (default: drop)\n"
+"  -a pemfile  use cert from pemfile when destination requests client certs\n"
+"  -b pemfile  use key from pemfile when destination requests client certs\n"
 #ifndef OPENSSL_NO_DH
 "  -g pemfile  use DH group params from pemfile (default: keyfiles or auto)\n"
 #define OPT_g "g:"
@@ -307,8 +309,8 @@ main(int argc, char *argv[])
 		natengine = NULL;
 	}
 
-	while ((ch = getopt(argc, argv, OPT_g OPT_G OPT_Z OPT_i "k:c:C:K:t:"
-	                    "OPs:r:R:e:Eu:m:j:p:l:L:S:F:M:dDVhW:w:q:")) != -1) {
+	while ((ch = getopt(argc, argv, OPT_g OPT_G OPT_Z OPT_i "k:c:C:K:t:OPa:"
+	                    "b:s:r:R:e:Eu:m:j:p:l:L:S:F:M:dDVhW:w:q:")) != -1) {
 		switch (ch) {
 			case 'c':
 				if (opts->cacrt)
@@ -428,6 +430,40 @@ main(int argc, char *argv[])
 				break;
 			case 'P':
 				opts->passthrough = 1;
+				break;
+			case 'a':
+				if (opts->clientcrt)
+					X509_free(opts->clientcrt);
+				opts->clientcrt = ssl_x509_load(optarg);
+				if (!opts->clientcrt) {
+					fprintf(stderr, "%s: error loading cli"
+					                "ent cert from '%s':\n",
+					                argv0, optarg);
+					if (errno) {
+						fprintf(stderr, "%s\n",
+						        strerror(errno));
+					} else {
+						ERR_print_errors_fp(stderr);
+					}
+					exit(EXIT_FAILURE);
+				}
+				break;
+			case 'b':
+				if (opts->clientkey)
+					EVP_PKEY_free(opts->clientkey);
+				opts->clientkey = ssl_key_load(optarg);
+				if (!opts->clientkey) {
+					fprintf(stderr, "%s: error loading cli"
+					                "ent key from '%s':\n",
+					                argv0, optarg);
+					if (errno) {
+						fprintf(stderr, "%s\n",
+						        strerror(errno));
+					} else {
+						ERR_print_errors_fp(stderr);
+					}
+					exit(EXIT_FAILURE);
+				}
 				break;
 #ifndef OPENSSL_NO_DH
 			case 'g':